Redundancy: Including both <curses.h> and <ncurses.h> is unnecessary. On Gentoo (and most systems), <ncurses.h> is the standard header for the ncurses library and includes everything you need (WINDOW, initscr, COLOR_RED, etc.). <curses.h> is an older, sometimes compatibility-focused header that might not be needed unless you’re targeting a very specific setup (e.g., non-ncurses curses implementations). Since you’re linking -lncurses, <ncurses.h> alone should suffice.
Fix: Drop #include <curses.h> unless you have a specific reason for it
(e.g., porting to a system without ncurses).

Changes to be committed:
	modified:   color.cpp

+++ color.cpp
@@ -21,7 +21,7 @@ auto main(int argc, char *argv[]) -> int
 }
 auto print_in_middle(WINDOW *win, int starty, int startx, int width, const char *string) -> void
 {	int length, x, y;
-	float temp;
+	int temp;
 
 	if(win == NULL) win = stdscr;
 	getyx(win, y, x);
@@ -30,11 +30,11 @@ auto print_in_middle(WINDOW *win, int starty, int startx, int width, const char 
 	if(width == 0) width = COLS; // terminal 80 -> can be different 
 
 	length = strlen(string);
-	temp = (int)((float)(width - length)/ (float)2);
+	temp = (width - length)/ 2;
 	x = startx + (int)temp;
 	mvwprintw(win, y, x, "%s", string);
 	refresh();
 }
 
-// compile : g++ -std=c++17 color.cpp -ltinfo -lncurses -o color
+// compile : g++ -std=c++23 color.cpp -ltinfo -lncurses -o color
 // execute : ./color
